Revert the commit that introduced S3 support
Even after using S3 for over 4 years, every time it's a pain in the ass
to support and get to run smoothly. Difficulty of creating new users,
horrible web interface, super-complex bucket policies, never-ending
problems with "Access Denied" errors and super slow uploads. That's
enough for me to drop the support in favor of something simple.

I'm reverting this change to use nginx & regular public/uploads folder,
which will not work on platforms like heroku (which I'm not using).
However, it should be possible to override Carrierwave settings in your
app if you really need to use S3.

This reverts commit 8b58aaba10ef57722f8789c2b61e6501729d3a74.